Attach plugin name
