+++ src/main/options.js
@@ -5,7 +5,7 @@ const getSupportInfo = require("../common/support").getSupportInfo;
 const normalizer = require("./options-normalizer");
 const loadPlugins = require("../common/load-plugins");
 const resolveParser = require("./parser").resolveParser;
-const getPrinter = require("./get-printer");
+const getPlugin = require("./get-plugin");
 
 const hiddenDefaults = {
   astFormat: "estree",
@@ -53,17 +53,18 @@ function normalize(options, opts) {
   rawOptions.astFormat = parser.astFormat;
   rawOptions.locEnd = parser.locEnd;
   rawOptions.locStart = parser.locStart;
-  rawOptions.printer = getPrinter(rawOptions);
+  const plugin = getPlugin(rawOptions);
+  rawOptions.printer = plugin.printers[rawOptions.astFormat];
 
   const pluginDefaults = supportOptions
     .filter(
       optionInfo =>
-        optionInfo.pluginDefaults && optionInfo.pluginDefaults[options.parser]
+        optionInfo.pluginDefaults && optionInfo.pluginDefaults[plugin.name]
     )
     .reduce(
       (reduced, optionInfo) =>
         Object.assign(reduced, {
-          [optionInfo.name]: optionInfo.pluginDefaults[options.parser]
+          [optionInfo.name]: optionInfo.pluginDefaults[plugin.name]
         }),
       {}
     );
